+++ src/main/java/blusunrize/immersiveengineering/api/shader/IShaderEffectFunction.java
@@ -4,11 +4,14 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
 
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+
 /**
  * @author BluSunrize - 14.05.2018
  */
 @FunctionalInterface
 public interface IShaderEffectFunction
 {
-	void execute(World world, ItemStack shader, ItemStack item, String shaderType, Vec3d pos);
+	void execute(@Nonnull World world, @Nonnull ItemStack shader, @Nullable ItemStack item, @Nonnull String shaderType, @Nonnull Vec3d pos, @Nullable Vec3d direction, float scale);
 }